pdf_writer

Module types

Types used by specific PDF structures.

Structs

AnnotationFlags

Bitflags describing various characteristics of annotations.

FontFlags

Bitflags describing various characteristics of fonts.

OutlineItemFlags

Bitflags describing the appearance of an outline item.

SystemInfo

Specifics about a character collection.

UnicodeCmap

A builder for a /ToUnicode character map stream.

Enums

ActionType

What kind of action to perform when clicking a link annotation.

AnnotationIcon

Possible icons for an annotation.

AnnotationType

Kind of the annotation to produce.

ArtifactAttachment

Where a layout Artifact is attached to the page edge.

ArtifactSubtype

The various subtypes of pagination artifacts.

ArtifactType

The various types of layout artifacts.

AttributeOwner

Owner of the attribute dictionary.

BlendMode

How to blend source and backdrop.

BlockAlign

The block alignment.

BorderType

The kind of line to draw on the border.

CidFontType

The subtype of a CID font.

ColorSpaceOperand

A color space operand to the CS or cs operator.

DeviceNSubtype

Type of n-dimensional color space.

Direction

Predominant reading order of text.

FieldRole

The kind of form control.

FieldState

Whether a check box or radio button is checked.

FontStretch

The width of a font’s glyphs.

FunctionShadingType

What kind of shading to use for a function-based shading.

HighlightEffect

Highlighting effect applied when a user holds the mouse button over an annotation.

InlineAlign

The inline alignment.

InterpolationOrder

How to interpolate between the samples in a function of the sampled type.

LayoutBorderStyle

Layout border style.

LineCapStyle

How to terminate lines.

LineJoinStyle

How to join lines at corners.

ListNumbering

The list numbering type.

MaskType

What property in the mask influences the target alpha.

NumberingStyle

The numbering style of a page label.

OutputIntentSubtype

The output intent subtype.

OverprintMode

How to behave when overprinting for colorants with the value zero.

PageLayout

How the viewer should lay out the pages in the document.

PageMode

Elements of the viewer chrome that should be visible when opening the document.

PaintType

Type of paint for a tiling pattern.

Placement

Placement of an element.

PostScriptOp

PostScript operators for use in Type 4 functions.

ProcSet

What procedure sets to send to a PostScript printer or other output device.

RenderingIntent

How the output device should aim to render colors.

RubyAlign

The alignment of a ruby annotation.

RubyPosition

The position of a ruby annotation.

SMaskInData

What to do with in-data mask information in JPXDecode images.

StructRole

Role the structure element fulfills in the document.

TabOrder

What order to tab through the annotations on a page.

TableHeaderScope

The scope of a table header cell.

TextAlign

The text alignment.

TextDecorationType

The text decoration type (over- and underlines).

TextRenderingMode

How to render text.

TilingType

How to adjust tile spacing.

TransitionAngle

From where to where a page transition plays.

TransitionStyle

A kind of page transition.

TrappingStatus

Whether a document has been adjusted with traps.

WritingMode

Writing direction.