pdf_writer::types

Enum StructRole

pub enum StructRole {

    Document,
    Part,
    Art,
    Sect,
    Div,
    BlockQuote,
    Caption,
    TOC,
    TOCI,
    Index,
    NonStruct,
    Private,
    P,
    H1,
    H2,
    H3,
    H4,
    H5,
    H6,
    L,
    LI,
    Lbl,
    LBody,
    Table,
    TR,
    TH,
    TD,
    THead,
    TBody,
    TFoot,
    Span,
    Quote,
    Note,
    Reference,
    BibEntry,
    Code,
    Link,
    Annot,
    Ruby,
    Warichu,
    RB,
    RT,
    RP,
    WT,
    WP,
    Figure,
    Formula,
    Form,

}

Role the structure element fulfills in the document.

These are the predefined standard roles. The writer may write their own and then provide a mapping. PDF 1.4+.

Variants

Document

The whole document.

Part

A part of a document that may contain multiple articles or sections.

Art

An article with largely self-contained content.

Sect

Section of a larger document.

Div

Generic subdivision.

BlockQuote

A paragraph-level quote.

Caption

An image or figure caption.

TOC

Table of contents.

TOCI

Item in the table of contents.

Index

Index of the key terms in the document.

NonStruct

Element only present for grouping purposes that shall not be exported.

Private

Element present only for use by the writer and associated products.

P

A paragraph

H1

First-level heading.

H2

Second-level heading.

H3

Third-level heading.

H4

Fourth-level heading.

H5

Fifth-level heading.

H6

Sixth-level heading.

L

A list.

LI

A list item.

Lbl

Label for a list item.

LBody

Description of the list item.

Table

A table.

TR

A table row.

TH

A table header cell.

TD

A table data cell.

THead

A table header row group.

TBody

A table data row group.

TFoot

A table footer row group.

Span

A generic inline element.

Quote

An inline quotation.

Note

A foot- or endnote.

Reference

A reference to elsewhere in the document.

BibEntry

A reference to an external document.

Code

Computer code.

Link

A link.

Annot

An association between an annotation and the content it belongs to. PDF 1.5+

Ruby

Ruby annotation for CJK text. PDF 1.5+

Warichu

Warichu annotation for CJK text. PDF 1.5+

RB

Base text of a Ruby annotation. PDF 1.5+

RT

Annotation text of a Ruby annotation. PDF 1.5+

RP

Punctuation of a Ruby annotation. PDF 1.5+

WT

Text of a Warichu annotation. PDF 1.5+

WP

Punctuation of a Warichu annotation. PDF 1.5+

Figure

Item of graphical content.

Formula

Mathematical formula.

Form

Form widget.