pdf_writer::writers

Struct ColorSpace

pub struct ColorSpace<'a> { /* private fields */ }

Writer for a color space.

This struct is created by Chunk::color_space, Chunk::color_space, ImageXObject::color_space, Separation::alternate_color_space and Group::color_space.

Implementations

impl ColorSpace<'_>

CIE-based color spaces.

pub fn cal_rgb( self, white_point: [f32; 3], black_point: Option<[f32; 3]>, gamma: Option<[f32; 3]>, matrix: Option<[f32; 9]>, )

Write a CalRGB color space.

pub fn cal_gray( self, white_point: [f32; 3], black_point: Option<[f32; 3]>, gamma: Option<f32>, )

Write a CalGray color space.

pub fn lab( self, white_point: [f32; 3], black_point: Option<[f32; 3]>, range: Option<[f32; 4]>, )

Write a Lab color space.

pub fn icc_based(self, stream: Ref)

Write an ICCBased color space.

The stream argument should be an indirect reference to an ICC profile stream.

impl ColorSpace<'_>

Common calibrated color spaces.

pub fn srgb(self)

Write a CalRGB color space approximating sRGB.

Use an ICC profile for more accurate results.

pub fn adobe_rgb(self)

Write a CalRGB color space approximating Adobe RGB.

Use an ICC profile for more accurate results.

pub fn display_p3(self)

Write a CalRGB color space approximating Display P3.

Use an ICC profile for more accurate results.

pub fn pro_photo(self)

Write a CalRGB color space approximating ProPhoto.

Use an ICC profile for more accurate results.

pub fn eci_rgb(self)

Write a CalRGB color space for ECI RGB v1.

pub fn ntsc(self)

Write a CalRGB color space for NTSC RGB.

pub fn pal(self)

Write a CalRGB color space for PAL/SECAM RGB.

pub fn d65_gray(self)

Write a CalGray color space for CIE D65 at a 2.2 gamma, equivalent to sRGB, Adobe RGB, Display P3, PAL, …

pub fn d50_gray(self, gamma: Option<f32>)

Write a CalGray color space for CIE D50 (horizon light). Set a 1.8 gamma for ProPhoto or ECI RGB equivalency, 2.2 is another common value.

pub fn c_gray(self)

Write a CalGray color space for CIE C (north sky daylight) at 2.2 gamma, equivalent to NTSC.

pub fn e_gray(self, gamma: Option<f32>)

Write a CalGray color space for CIE E (equal emission). Common gamma values include 1.8 or 2.2.

impl ColorSpace<'_>

Device color spaces.

pub fn device_rgb(self)

Write a DeviceRGB color space.

pub fn device_cmyk(self)

Write a DeviceCMYK color space.

pub fn device_gray(self)

Write a DeviceGray color space.

impl<'a> ColorSpace<'a>

Special color spaces.

pub fn separation(self, color_name: Name<'_>) -> Separation<'a>

Start writing a Separation color space. PDF 1.2+.

The color_name argument is the name of the colorant that will be used by the printer.

pub fn device_n<'n>( self, names: impl IntoIterator<Item = Name<'n>>, ) -> DeviceN<'a>

Write a DeviceN color space. PDF 1.3+.

The names argument contains the N names of the colorants for the respective components.

pub fn indexed(self, base: Name<'_>, hival: i32, lookup: &[u8])

Write an Indexed color space. PDF 1.2+.

The length of the lookup slice must be the product of the dimensions of the base color space and (hival + 1) and hival shall be at most 255.

pub fn pattern(self, base: Name<'_>)

Write a Pattern color space for uncolored patterns. PDF 1.2+.

The base attribute is the color space in which the pattern’s tint color is specified upon use.