pdf_writer::writers

Struct FontDescriptor

pub struct FontDescriptor<'a> { /* private fields */ }

Writer for a font descriptor dictionary.

This struct is created by Chunk::font_descriptor.

Implementations

impl<'a> FontDescriptor<'a>

pub fn name(&mut self, name: Name<'_>) -> &mut Self

Write the /FontName attribute. Required.

pub fn family(&mut self, family: Str<'_>) -> &mut Self

Write the /FontFamily attribute. Recommended for Type 3 fonts in Tagged PDFs. PDF 1.5+.

pub fn stretch(&mut self, stretch: FontStretch) -> &mut Self

Write the /FontStretch attribute. Recommended for Type 3 fonts in Tagged PDFs. PDF 1.5+.

pub fn weight(&mut self, weight: u16) -> &mut Self

Write the /FontWeight attribute. Should be between 100 (lightest) and 900 (heaviest), 400 is normal weight, 700 is bold. Recommended for Type 3 fonts in Tagged PDFs. PDF 1.5+.

pub fn flags(&mut self, flags: FontFlags) -> &mut Self

Write the /Flags attribute. Required.

pub fn bbox(&mut self, bbox: Rect) -> &mut Self

Write the /FontBBox attribute. Required, except for Type 3 fonts.

pub fn italic_angle(&mut self, angle: f32) -> &mut Self

Write the /ItalicAngle attribute. Required.

pub fn ascent(&mut self, ascent: f32) -> &mut Self

Write the /Ascent attribute. Required.

pub fn descent(&mut self, descent: f32) -> &mut Self

Write the /Descent attribute. Required.

pub fn leading(&mut self, leading: f32) -> &mut Self

Write the /Leading attribute.

pub fn cap_height(&mut self, cap_height: f32) -> &mut Self

Write the /CapHeight attribute. Required for fonts with Latin characters, except for Type 3 fonts.

pub fn x_height(&mut self, x_height: f32) -> &mut Self

Write the /XHeight attribute.

pub fn stem_v(&mut self, stem_v: f32) -> &mut Self

Write the /StemV attribute. Required, except for Type 3 fonts.

pub fn stem_h(&mut self, stem_h: f32) -> &mut Self

Write the /StemH attribute.

pub fn avg_width(&mut self, avg_width: f32) -> &mut Self

Write the /AvgWidth attribute.

pub fn max_width(&mut self, max_width: f32) -> &mut Self

Write the /MaxWidth attribute.

pub fn missing_width(&mut self, missing_width: f32) -> &mut Self

Write the /MissingWidth attribute.

pub fn font_file(&mut self, id: Ref) -> &mut Self

Write the /FontFile attribute, referecing Type 1 font data.

pub fn font_file2(&mut self, id: Ref) -> &mut Self

Write the /FontFile2 attribute, referencing TrueType font data. PDF 1.1+.

pub fn font_file3(&mut self, id: Ref) -> &mut Self

Write the /FontFile3 attribute, referencing CFF font data. PDF 1.2+ or PDF 1.3+ for CID-keyed fonts.

pub fn char_set(&mut self, names: Str<'_>) -> &mut Self

Write the /CharSet attribute, encoding the character names of a font subset as a string. This is only relevant for Type 1 fonts. PDF 1.1+.