pdf_writer::writers

Struct LayoutAttributes

pub struct LayoutAttributes<'a> { /* private fields */ }

Writer for an layout attributes dictionary. PDF 1.4+

This struct is created by Attributes::layout.

Implementations

impl<'a> LayoutAttributes<'a>

General layout attributes.

pub fn placement(&mut self, placement: Placement) -> &mut Self

Write the /Placement attribute.

pub fn writing_mode(&mut self, mode: WritingMode) -> &mut Self

Write the /WritingMode attribute to set the writing direction.

pub fn background_color(&mut self, color: [f32; 3]) -> &mut Self

Write the /BackgroundColor attribute to set the background color in RGB between 0 and 1. PDF 1.5+

pub fn border_color(&mut self, color: [f32; 3]) -> &mut Self

Write the /BorderColor attribute.

pub fn border_style(&mut self, style: [LayoutBorderStyle; 4]) -> &mut Self

Write the /BorderStyle attribute.

pub fn border_thickness(&mut self, thickness: [f32; 4]) -> &mut Self

Write the /BorderThickness attribute.

pub fn padding(&mut self, padding: [f32; 4]) -> &mut Self

Write the /Padding attribute.

pub fn color(&mut self, color: [f32; 3]) -> &mut Self

Write the /Color attribute.

impl LayoutAttributes<'_>

Block level elements.

pub fn space_before(&mut self, space: f32) -> &mut Self

Write the /SpaceBefore attribute.

pub fn space_after(&mut self, space: f32) -> &mut Self

Write the /SpaceAfter attribute.

pub fn start_indent(&mut self, indent: f32) -> &mut Self

Write the /StartIndent attribute.

pub fn end_indent(&mut self, indent: f32) -> &mut Self

Write the /EndIndent attribute.

pub fn text_indent(&mut self, indent: f32) -> &mut Self

Write the /TextIndent attribute.

pub fn text_align(&mut self, align: TextAlign) -> &mut Self

Write the /TextAlign attribute.

pub fn width(&mut self, width: f32) -> &mut Self

Write the /Width attribute for table row groups and illustrative elements. No intrinsic height will be assumed if left empty.

pub fn height(&mut self, height: f32) -> &mut Self

Write the /Height attribute for table row groups and illustrative elements. No intrinsic height will be assumed if left empty.

impl LayoutAttributes<'_>

Illustration elements.

pub fn bbox(&mut self, bbox: Rect) -> &mut Self

Write the /BBox attribute.

impl LayoutAttributes<'_>

Table header and data.

pub fn block_align(&mut self, align: BlockAlign) -> &mut Self

Write the /BlockAlign attribute.

pub fn inline_align(&mut self, align: InlineAlign) -> &mut Self

Write the /InlineAlign attribute.

pub fn table_border_style(&mut self, style: [LayoutBorderStyle; 4]) -> &mut Self

Write the /TBorderStyle attribute. PDF 1.5+.

pub fn table_padding(&mut self, padding: f32) -> &mut Self

Write the /TPadding attribute. PDF 1.5+.

impl<'a> LayoutAttributes<'a>

Grouping elements.

pub fn column_count(&mut self, count: i32) -> &mut Self

Write the /ColumnCount attribute. PDF 1.6+.

pub fn column_widths(&mut self) -> TypedArray<'_, f32>

Start writing the /ColumnWidths array. The last number in the array is used for all extra columns. PDF 1.6+.

pub fn column_gap(&mut self) -> TypedArray<'_, f32>

Start writing the /ColumnGap array. The last number in the array is used for all extra columns. PDF 1.6+.

impl LayoutAttributes<'_>

Inline elements.

pub fn line_height(&mut self, height: LineHeight) -> &mut Self

Write the /LineHeight attribute.

pub fn baseline_shift(&mut self, shift: f32) -> &mut Self

Write the /BaselineShift attribute.

pub fn text_decoration_type( &mut self, decoration: TextDecorationType, ) -> &mut Self

Write the /TextDecorationType attribute. PDF 1.5+.

pub fn text_decoration_color(&mut self, color: [f32; 3]) -> &mut Self

Write the /TextDecorationColor attribute in RGB. PDF 1.5+.

pub fn text_decoration_thickness(&mut self, thickness: f32) -> &mut Self

Write the /TextDecorationThickness attribute. PDF 1.5+.

impl LayoutAttributes<'_>

Vertical Text.

pub fn glyph_orientation_vertical(&mut self, angle: f32) -> &mut Self

Write the /GlyphOrientationVertical attribute as an angle between -90 and 360 in multiples of 90. PDF 1.5+.

impl LayoutAttributes<'_>

Ruby annotations.

pub fn ruby_align(&mut self, align: RubyAlign) -> &mut Self

Write the /RubyAlign attribute. PDF 1.5+.

pub fn ruby_position(&mut self, position: RubyPosition) -> &mut Self

Write the /RubyPosition attribute. PDF 1.5+.