pdf_writer::writers

Struct Page

pub struct Page<'a> { /* private fields */ }

Writer for a page dictionary.

This struct is created by Chunk::page.

Implementations

impl<'a> Page<'a>

pub fn parent(&mut self, parent: Ref) -> &mut Self

Write the /Parent attribute. Required.

pub fn last_modified(&mut self, date: Date) -> &mut Self

Write the /LastModified attribute. PDF 1.3+.

pub fn media_box(&mut self, rect: Rect) -> &mut Self

Write the /MediaBox attribute. This is the size of the physical medium the page gets printed onto.

pub fn crop_box(&mut self, rect: Rect) -> &mut Self

Write the /CropBox attribute. This is the size of the area within which content is visible.

pub fn bleed_box(&mut self, rect: Rect) -> &mut Self

Write the /BleedBox attribute. This is the size of the area within which content is visible in a print production environment. Most production-aiding marks should be outside of this box. PDF 1.3+.

pub fn trim_box(&mut self, rect: Rect) -> &mut Self

Write the /TrimBox attribute. This is the size of the produced document after trimming is applied. PDF 1.3+.

pub fn art_box(&mut self, rect: Rect) -> &mut Self

Write the /ArtBox attribute. This is the area that another program importing this file should use. PDF 1.3+.

pub fn resources(&mut self) -> Resources<'_>

Start writing the /Resources dictionary.

pub fn contents(&mut self, id: Ref) -> &mut Self

Write the /Contents attribute as reference to a single content stream.

Such a content stream can be created using the Content builder and written to the file using Chunk::stream.

pub fn contents_array( &mut self, ids: impl IntoIterator<Item = Ref>, ) -> &mut Self

Write the /Contents attribute as an array.

pub fn rotate(&mut self, degrees: i32) -> &mut Self

Write the /Rotate attribute. This is the number of degrees the page should be rotated clockwise when displayed. This should be a multiple of 90.

pub fn group(&mut self) -> Group<'_>

Start writing the /Group dictionary to set the transparency settings for the page. PDF 1.4+.

pub fn thumbnail(&mut self, id: Ref) -> &mut Self

Write the /Thumb attribute to set an image as the page thumbnail. Must be RGB, Grayscale, or an indexed color space based on the former two.

pub fn duration(&mut self, seconds: f32) -> &mut Self

Write the /Dur attribute. This is the amount of seconds the page should be displayed before advancing to the next one. PDF 1.1+.

pub fn transition(&mut self) -> Transition<'_>

Start writing the /Trans dictionary. This sets a transition effect for advancing to the next page. PDF 1.1+.

pub fn annotations(&mut self) -> TypedArray<'_, Annotation<'_>>

Start writing the /Annots (annotations) array.โ

pub fn struct_parents(&mut self, key: i32) -> &mut Self

Write the /StructParents attribute to indicate the structure tree elements the contents of this XObject may belong to. PDF 1.3+.

pub fn tab_order(&mut self, order: TabOrder) -> &mut Self

Write the /Tabs attribute. This specifies the order in which the annotations should be focused by hitting tab. PDF 1.5+.

pub fn user_unit(&mut self, value: f32) -> &mut Self

Write the /UserUnit attribute. This sets how large a user space unit is in printerโs points (1/72 inch). This defaults to 1.0. PDF 1.6+.

pub fn metadata(&mut self, id: Ref) -> &mut Self

Write the /Metadata attribute to specify the pageโs metadata. PDF 1.4+.

The reference shall point to a metadata stream.