pdf_writer::writers

Struct StructTreeRoot

pub struct StructTreeRoot<'a> { /* private fields */ }

Writer for a structure tree root dictionary. PDF 1.3+

This struct is created by Catalog::struct_tree_root.

Implementations

impl<'a> StructTreeRoot<'a>

pub fn child(&mut self, id: Ref) -> &mut Self

Write the /K attribute to reference the immediate child of this element.

pub fn children(&mut self) -> TypedArray<'_, Ref>

Start writing the /K attribute to reference the immediate children of this element.

pub fn id_tree(&mut self) -> NameTree<'_, Ref>

Start writing the /IDTree attribute to map element identifiers to their corresponding structure element objects. Required if any elements have element identifiers.

pub fn parent_tree(&mut self) -> NumberTree<'_, Ref>

Start writing the /ParentTree attribute to maps structure elements to the content items they belong to. Required if any structure elements contain content items.

pub fn parent_tree_next_key(&mut self, key: i32) -> &mut Self

Write the /ParentTreeNextKey attribute to specify the next available key for the /ParentTree dictionary.

pub fn role_map(&mut self) -> RoleMap<'_>

Start writing the /RoleMap attribute to map structure element names to their approximate equivalents from the standard set of types. PDF 1.4+.

pub fn class_map(&mut self) -> ClassMap<'_>

Start writing the /ClassMap attribute to map objects designating attribute classes to their corresponding attribute objects or arrays thereof.