pdf_writer::writers

Struct Action

pub struct Action<'a> { /* private fields */ }

Writer for an action dictionary.

This struct is created by Annotation::action.

Implementations

impl<'a> Action<'a>

pub fn action_type(&mut self, kind: ActionType) -> &mut Self

Write the /S attribute to set the action type.

pub fn destination(&mut self) -> Destination<'_>

Start writing the /D attribute to set the destination of this GoTo-type action.

pub fn destination_named(&mut self, name: Name<'_>) -> &mut Self

Write the /D attribute to set the destination of this GoTo-type action to a named destination.

pub fn file_spec(&mut self) -> FileSpec<'_>

Start writing the /F attribute, setting which file to go to or which application to launch.

pub fn new_window(&mut self, new: bool) -> &mut Self

Write the /NewWindow attribute to set whether this remote GoTo action should open the referenced destination in another window.

pub fn uri(&mut self, uri: Str<'_>) -> &mut Self

Write the /URI attribute to set where this link action goes.

pub fn is_map(&mut self, map: bool) -> &mut Self

Write the /IsMap attribute to set if the click position of the userโs cursor inside the link rectangle should be appended to the referenced URI as a query parameter.